ucx: relocate GEX_EVENT_NOW stall outward
This commit relocates the stall for local completion of a Long payoad
when `lc_opt == GEX_EVENT_NOW`, with the new placement closely aligned
with the nearest equivalent logic in ibv-conduit.

Previously the stall took place between the payload and header
injections, which gave the incorrect impression (see bug 4155,
comments 2 and 3) that it was related to ordering between those two
network sends.  If it was for that purpose, then it would not have
been sufficient to provide the *remote* completion needed for that
purpose.

This change will simplify resolution of bug 4155, as well as leading
to an additional simplification due to the `counter` parameter having
now become unused.