ucx: fix RequestSysShort initiation
This commit repairs an incorrect call to `gasnetc_am_reqrep_inner()`
in `gasnetc_RequestSysShort()`.  Previously the `is_request` boolean
parameter was `0`, when `1` is the proper value for a request.

The defect appears to have originated before the `is_sync` parameter
was split from `is_request` (to prevent System-level request injection
from being "sync"), but it was not updated when that split occurred.