ucx: correct exit coordination AMs
This commit removes incorrect counter manipulations in the logic used
to coordinate some non-collective exit cases.  In particular, in
`testexit` cases 3, 8, 10, 11, 12 and 13 are known to have changed
from exit on timer expiration to graceful exit as a result of this
commit.

When commit 29ad84a5 ported ibv-conduit's graceful exit handling, it
utilized the `gasnetc_{Request,Reply}SysShort() functions` added in
commit 770c48a1.  However, there is a subtle difference in that
implementation, relative to ibv-conduit's, which was not taken into
account.  Specifically, the ibv-conduit code would pre-increment a
counter which the `SysShort` AM injections would balance when the
AM injection was complete (as opposed to queued), in order to prevent
shutdown from destroying a queue prior to such injection.  However,
the use of callbacks in the ucx-conduit version of the `SysShort`
implementation *provides* the initial increment, making the
pre-increment incorrect such that the counter would never balance
in certain cases.  The PSHM path is also updated to match.

This went unnoticed/unreported because the relevant code does not run
in normal execution *and* the suicide timer would eventually allow the
exit to complete (just with one line of extra noise).