ucx: correct handling for sync LC for Long
This commit adopts logic from ibv-conduit to simplify the code paths
when using put-sync-send for Long payloads (to deal with bug 4155).
In particular, since one stalls for remote completion in this case
there is no need to track local completion.  However, this requires
logic at exit from the AM injection to "cleanup" when local
completions was synchronous for this reason.

This resolves a "TODO" item from the original fix for bug 4155.