ucx: replace a 2-case 'switch' with an 'if'
This commit acts on the observation that a `switch` statement had been
written with a `case` for Short AMs that was known to be unreachable
(as acknowledged in a comment: "only to silence the warning").  This
is replaced with an `if` which should be at least as simple for the
human reader, as well as hopefully generating better code (certainly
not worse).  An assertion is added to fail if the `else` block is ever
reached in error.