ucx: improve control over bug 4155 fix
This commit implements and documents a new environment variable,
`GASNET_UCX_AM_ORDERED_TLS`, which provides a manual override for
the choice between "put+send" and "put+sync+send" (for ordered and
unordered cases, respectively) algorithms for AM Long payloads.

Additionally, the default is made more intelligent.  Where previously
the choice exactly followed `GASNET_PSHM`, the new default is based on
the presence of exactly one nbrhd (aka supernode) per node.  This is
exactly the case in which GASNet (via PSHM or true loopback) is used
for all intra-node AMs.  This corrects the behavior for multi-nbrhd
hosts (as via `GASNET_SUPERNODE_MAXSIZE`) which previously could have
been incorrect for cross-nbrhd comms via UCX shared-memory transports.

The revised default algorithm selection also for allows use of the
faster in-order algorithm in a non-PSHM build when communicating
off-node from a process which has no on-node peers (IOW, the 1-ppn
case is considered to be one nbrhd).  This is an important case for
clients such as Legion and Chapel which disable PSHM precisely because
they run 1-ppn as the normal case.