Factor gasnete_eop_new_alc()
This commit adds a new `gasnete_eop_new_alc()` internal function to
factor the idiom in which the conduit creates a eop which is used
only for LC of AM payload injection.  This idiom is now common to
ibv and ucx conduit and is a natural one to expect other conduits
to make use of.