testcontend: prevent division by zero
This commit makes two minor changes to testcontend's report generation
to avoid division by zero.  The first is to convert ticks to ns rather
than us.  This avoids a zero value which was occurring on some NDEBUG
builds of smp-conduit simply because the default 50 rounds of 8-byte
put (or get) complete in less than 1us.  The second is an explicit
check for zero, since too coarse a resolution for the underlying
timer might still report a 0ns result (zero ticks).