Update comments to reflect fix for bug 4278
With the resolution of bug 4278, the conduit-independent RDMADISSEM
barrier no longer contains any "complications due to thread-specific
events".  So, this commit updates comments in the analogous aries- and
ibv-specific dissemination barriers to no longer list that as a "key
difference".  Instead they mention the lack of overheads associated
with the fully general eop or iop (though only the latter is used).