Allow leak of implicit iop on thread exit
This commit replaces an assertion with an `if` in the logic for thread
destruction to ensure that a thread which exits with outstanding NBI
operations will leak memory rather than an assertion failure.

Additionally, `gasnete_iop_isdone()` is now defined unconditionally,
where is was previously debug-specific (when it was only used in an
assertion).

Future work described in bug 4287 should resolve the leak, fwiw.