Improve performance of RDMADISSEM barrier
This commit resolves the issues identified in bug 4278 by eliminating
the use of explicit events and NBI access regions in the RDMADISSEM
barrier implementation, in favor of "bleeding" small puts into the
client's then-current iop.

The main impact of this change is to avoid turning every put in a
multi-threaded build into blocking.  However, it also removes the sync
of events at barrier completion, reducing the barrier latency for SEQ
builds as well.