ibv: fix bugs 4148 and 4150
This commit repairs the injection logic for RMA Put targeting device
memory to pass the correct completion counter/callback for the cases
in which LC is not tracked separately.  This includes all use of
`gex_RMA_PutBlocking()` and of uses `gex_RMA_Put{NB,NBI}()` with
`lc_opt == GEX_EVENT_DEFER`.

Prior to this correction, blocking puts could return immediately
without stalling for either local or remote completion, and
non-blocking calls with `GEX_EVENT_DEFER` would return with state
consistent with both local and remote completion having occurred,
regardless of whether either actually had.  These flaws caused the
erroneous behaviors reported as bug 4148 and bug 4150.

This commit is essentially a critically important fixup to
commit 2b7377c218efea7e4467c031a1df96087e867de6
   "ibv: gex_RMA_*() support for CUDA UVA memory"

While this commit definitely fixes a real logic error, it is unclear
at this point in time whether there are additional issues (such as
with the GDR implementation or the GPU memory model) which may also
contribute to the behaviors reported as bugs 4148 and 4150.
Therefore, we are deferring updates to ChangeLong and memory_kinds.md
until this change has seen additional testing post-merge.