ucx: AM header reductions, part 1
This commit reduces the size of the on-the-wire AM header from 36
bytes (40 in a debug build) to 24.  With padding for 8-byte payload
alignment, this translates to a 16-byte increase in LUB Medium (from
3992 to 4008 for the default) for all builds.

 + The `magic` field was set but never read.
 + The `dst` was used only on the sender, not on the wire.
 + Only 2 bits are needed for category (aka `am_type`) but it was
   previously occupying 32 bits, positioned between two single-byte
   fields creating several bytes of *avoidable* padding.

The Boolean `is_packed` member has no net impact of the size of the
header (residing in alignment padding).  However, since it is
currently read from only in EVERYTHING segment mode, its definition
and initialization are now omitted for FAST and LARGE.

No effort has yet been made to omit the Long-specific `dst_addr` from
the wire when sending Short or Medium AMs, nor to omit `payload_size`
from Short AMs.  These changes are the subject of upcoming work.

I suspect it should be possible to eliminate the `src`, since the
recipient should be able to recover that from the tag-matching
send/recv APIs.  Additionally, `payload_size` is likely recoverable
from the `length` available to the recipient via the tag-matching
send/recv APIs.

It appears to me that `size` actually serves no purpose at the
recipient other than in assertions.  As computed now, it should be
identical to the `length` I believe should be available to the
recipient via the tag-matching send/recv APIs.  At a minimum, the
`size` and `payload_size` should be each be computable from the other
using `am_type` and `num_args`.  However, that computation is
currently the *subject* of the use of `size` in assertions.  So, I am
not in a rush to remove this field.

Finally, this commit removes a redundant assignment in
`gasnetc_send_req()`, which was duplicating the variable's
initialization (changed in this commit).