ucx: AM header reductions, part 2
This commit varies the AM header on the wire, reducing the size of the
header for Short and Medium by not sending unused fields.  This yields
an additional 8 bytes of LUB Medium payload, and shortens Shorts by 12
bytes.

The macros in `gasnet_core.h` relating to AM payload and header sizes
are revised significantly.  Most significant is the elimination of a
`gasnetc_AMHeaderSize()` function call in the max payload queries.
That was previously a work-around for the fact that the AM header
struct was not in scope in this header to apply `sizeof()` to.  This
commit defines literal integer values for the Short, Medium and Long
header sizes and (critically) adds `gasneti_static_assert()` calls to
ensure they are never out-of-date.

Several assertions related to header and message size have been
updated to be correct for the new variable header size, and/or to use
`gasneti_assert_uint()`.  The reception of a Long previously lacked
assertions checking the total message size, but these have been added.

The `size` field is only used at the recipient in the assertions
mentioned above, and the sender does not consume `size` at all.  So,
it remains a candidate for future removal.  However, that would come
at the expense of the assertions, which were critical to getting the
new logic right.  So, future work remains to either replace `size` by
a message length obtained from the tag matching APIs, or to make
`size` a debug-only field.