Implement a conduit-independent "aop" facility
This commit extends the conduit-independent OP interface with routines
with a `gasneti_aop_` prefix to manage access region ops which can be
suspended/resumed.  This includes documentation in
`gasnet_internal.h`, implementation in `gasnet_event.h` and some basic
tests have been added to `op_test() in `gasnet_diagnostic.c`.

As part of this work, the implementation of NBI access regions has
been recast in terms of an internal parallel to the aop, to retain a
single instance of the common implementation details.  Specifically, a
`gasnete_aop_*()` family of inline functions is introduced to as the
foundation for both `gasnete_{begin,end}_nbi_accessregion()` and the
new `gasneti_aop_*()` interfaces.