Implement new GASNET_HOST_DETECT env var
This commit implements and documents a new `GASNET_HOST_DETECT`
environment variable to select the source for the identifier used to
build the "nodemap" used by PSHM to construct nbrhds (aka supernodes)
and to implement `gex_System_QueryHostInfo()`.

The implementation includes a configure option, `--with-host-detect`,
to specify a site default for the environment variable.

The default behaviors of all conduits remain unchanged, including
those of aries and udp which provide their own identifiers.  The
ability to use `gethostid()` for udp-conduit via the legacy variable
`GASNET_USE_GETHOSTID` is retained, but deprecated in the
documentation and via a warning at runtime.

Future work is planned to include an "greedy" default to use all the
available identifiers to identify the largest (most optimistic) nbrhd
possible, but to also provide the means to identify which identifiers
are "flawed" (one-to-many or many-to-one flaws are possible).  Future
work may also change the aries and/or udp defaults.

This resolves
Bug 4297 - incorrect nbrhd construction for some multi-homed hosts