Make gasneti_check_jobrank() an inline function
This commit converts `gasneti_check_jobrank()` from a macro to an
inline function.  This allows it to appear in an expression context,
where the previous macro expansion using `gasneti_assert_uint()`
could not.

This is used in expression context in `gasneti_in_auxsegment()` for
SEGMENT_EVERYTHING, which has not yet had a caller.