fixup: Implement new GASNETI_NBRHD_{MAPPED,AUXSEG}*()
This commit corrects assertion login to not verify existence of a
bound primordial segment prior to ensuring the target *is* a
primordial segment.

Additionally, the case of `GASNETI_SUPPORTS_OUTOFSEGMENT_PUTGET` is
added to `GASNET_SEGMENT_FAST` as one in which the assertions must be
skipped.  In particular, because this mode allows for RMA operations
targeting out-of-segment remote addresses, it is valid to reach the
query function without the existence of a bound segment.