Stubs for GASNETC_SEGMENT_DESTROY_HOOK
This commit provides the "stubs" for a conduit-specific hook to be
called when a segment is destroyed.

Currently there is no *actual* caller.  While the internal function
`gasneti_segmentDestroy()` is written with the *possibility* of
calling this hook, it's sole planned caller in the near-term will
disable such calls by passing a zero value for the
`create_hook_succeeded` parameter.

This hook will become relevant when `gex_Segment_Destroy()` is
implemented.  In preparation for that time, comments have been added
to the existing segment create hooks to record their status with
respect to the design in which destroy hooks do not run following a
failed create hook.