vis: discontinue use of _gex_RMA_*() entry points
This commit replaces calls within the VIS code to use `gex_RMA_*()`
entry points rather than the underscore-prefixed ones.  This will
enable planned improvements to the implementation of the RMA
interfaces which require breaking these internal ones.

The only key difference between the two sets of interfaces is the
handling of `THREAD_FARG` goop.  However, the recent improvements
to that subsystem, there is no reason to believe that there is a
meaningful difference between `_PASS` (with `_gex_RMA*()`) versus
`_GET` (with `gex_RMA*()`).