Warning suppression for --with-hip-platform=nvcc
This commit adds additional warning suppression flags for warnings
from `hip_runtime.h` in `__HIP_PLATFORM_NVCC__` mode when compiling
`kinds_hip.c` and `testhip.c`.  These flags are *only* applied when
compiling code calling the HIP library in `__HIP_PLATFORM_NVCC__`
mode.

Additionally, this commit modifies command lines in
`tests/Makefile.in` and `other/Makefile-conduit.mak.in` to place the
kind-specific CFLAGS *after* the generic ones.  This is necessary in
(at least) the case of `clang` which (unlike `gcc`) is sensitive to
the relative order of `-Wall` and `-Wno-*` on the command line.

This work is sufficient to compile `libgasnet-ibv-*.q` and `testhip`
without warnings with recent GNU, Clang, Intel, PGI and NVHPC
compilers on Dirac.