HIP: use hipGetErrorName()
This commit uses `hipGetErrorName()` to provide HIP error messages of
similar nature and quality as their CUDA_UVA counterparts.  While this
function is missing from the CUDA/HIP quick reference chart I had
bookmarked, that is probably because the signatures differ.  This
function certainly exists and has for some time.