testhip: validate GASNet segment allocation
This commit adds logic to `testhip` to validate that GASNet-allocated
device segments reside on the device passed when creating the Kind
(when multiple devices are available).  With `-gasnet-seg`, this is
sufficient to reproduce the issue reported in bug 4318 based on source
inspection.

Additionally, the new logic checks that `gex_Segment_Create()` does
not change the calling thread's current HIP device.