Update use of HIP platform identifiers
This commit resolves bug 4317 by ensuring that GASNet defines both the
old and new `__HIP_PLATFORM_*` identifiers.  This is best practice for
code not otherwise requiring a sufficiently recent ROCm version.

The logic in `testhip.c` is updated to test for only the new
identifier, since we control its setting in `gasnet_mk.h`.

Additionally, the internal identifiers and maintainer-facing configure
option have been updated to favor the new names to reduce the
likelihood of future confusion.  However, the configure option still
permits the old values, to avoid a need for coordinated updates to
pushbuild and the private builds of GASNet-EX developers.