GASNet-EX.txt: Strengthen gex_TM_Create(LOCAL_SCRATCH) to guarantee an output barrier
We agreed on the 2021-09-13 this is provided by the current
implementation of LOCAL_SCRATCH and is unlikely to ever be relaxed.

We hope to eventually deploy a less synchronous team create call,
especially once scratch space is managed transparently by the
implementation, but this would not affect operation for legacy callers
passing the LOCAL_SCRATCH flag.