ucx: fix bug 4328: issue in gasnetc_poll_sndrcv()
As described in bug 4328, the while loop in `gasnetc_poll_sndrcv()`
may rarely erroneous process a not-ready entry from the head of the
recv queue if the prior head tests as ready a second time.  This
commit resolves the problem by converting the `while` to an `if`, such
that there is no "second time".

The alternative of correcting the `while` by re-reading the head of
the queue was rejected, because the result theoretically might never
terminate.

Additionally, a `gasneti_assume()` has been added to let the compiler
know the request pointer read before the `if` and the one popped
within it are the same.