Fix bug 4330 (bad HIDDEN_AM_CONCURRENCY_LEVEL)
This commit fixes bug 4330 "ibv conduit incorrectly implements
HIDDEN_AM_CONCURRENCY_LEVEL" to honor the "iff" portion of the
specification which prohibits defining to `1` when there is not
an actual possibility of AM handler execution in conduit-owned
threads.

The top-level logic in `gasnetex.h` now defers to the conduit.
Meanwhile, the value set by ibv-conduit is now tied to the enable of
the recv thread, distinct from `GASNETI_CONDUIT_THREADS`.