ucx: resolve bug 4298
This commit resolves Bug 4298 "ucx persona_example hang w/o PSHM" by
ensuring that we fully progress the conduit (subject to handler
context restriction) when spinning for an AM send buffer.

This commit also adds polling of PSHM to `gasnetc_poll_snd()` in FAST
and LARGE segment modes (where it was previously missing) and removes
it from EVERYTHING mode (where PSHM is not supported).  This is
somewhat orthogonal to the bug report (which is non-PSHM), but in the
spirit of fully/correctly polling when stalled for resources.