aries: update design docs
This commit makes some significant updates to `aries-conduit/README`
to more accurately describe the mechanism(s) behind the Core API
implementation.

Presently the document lacks any description of the bitmap used at the
initiator to manage buffers at the target.  Additionally, the section
"Event Processing" is still out-of-date with respect to things like
the Gets in the AM rendezvous protocol.