ibv: error reporting improvements
This commit greatly improved the error messages generated when an ibv-level
error is signaled in the CQ.  As illustrated below, the information content has
improved.  Additionally, use of `gasneti_console_message()` ensures a more
uniform formatting and preservation of the messages in the trace file (if any).

Before, a typical output might look like the following:
```
@ 0> snd status=10 opcode=1 dst_node=1 dst_qp=0
@ 0> - rcv CQ contains impossibly large WCE count with status 10
*** FATAL ERROR (proc 0): in gasnetc_snd_reap() at [redacted]/gasnet/ibv-conduit/gasnet_core_sndrcv.c:843: aborting on reap of failed send
```

With this patch, the NDEBUG output is:
```
*** ERROR (proc 0): rcv status=5(Work Request Flushed Error) vendor_err=0xf9 qp_num=0x59e hca=mlx4_0
*** ERROR (proc 0): snd status=10(remote access error) vendor_err=0x88 qp_num=0x59e hca=mlx4_0 op=PUT_INLINE(RDMA_WRITE) dest=(proc:1, qpi:0)
*** FATAL ERROR (proc 0): in gasnetc_rcv_reap() at [redacted]/gasnet/ibv-conduit/gasnet_core_sndrcv.c:1237: aborting on reap of failed recv
```

Note that the "impossibly large" message is gone, replaced by a "Work Request
Flushed Error" line (which might occur in either order with respect to the
`snd` error line).  In the past, if the flushed rcv CQ was polled before the
snd CQ (as has occurred here), then the *real* error (a "remote access
error" in this case) was not reported.

The old "impossibly large" message turned out to be due to a subtle error in
the prior logic, which was interpreting `0` as success, rather than its actual
meaning of "CQ is empty" (successful poll of the CQ returns the number of CQEs
reaped, always `1` in our usage).  This error was preventing the reporting of
the error on the "other" CQ, which has always been intended.

The new DEBUG output is even richer, providing some detail of failing ibv-level
operations on the snd CQ (such as flags, addresses and length in this case):

```
*** ERROR (proc 0): rcv status=5(Work Request Flushed Error) vendor_err=0xf9 qp_num=0x6e4 hca=mlx4_0
*** ERROR (proc 0): snd status=10(remote access error) vendor_err=0x88 qp_num=0x6e4 hca=mlx4_0 op=PUT_INLINE(RDMA_WRITE) dest=(proc:1, qpi:0) flags=INLINE loc_addr=0x00007ffd 721d5748 rem_addr=0x00007fa8 4cc80000 length=4
*** FATAL ERROR (proc 0): in gasnetc_rcv_reap() at [redacted]/gasnet/ibv-conduit/gasnet_core_sndrcv.c:1237: aborting on reap of failed recv
```