testexit: two bug fixes for use of ReplyLong
This commit corrects `testexit.c` for the case of more than two
processes by ensuring that Long Replies use the segment of the
correct sender.

Additionally, the segment is now split into two halves to ensure that
the loopback case does not issue a Long Reply that has overlapping
source and destination.