ucx: exit improvements (including for bug 4360)
This commit makes the following improvements to the logic implementing
`gasnet_exit()` in ucx-conduit:

1. All fixed-duration `alarm()` calls in the exit code have been
subsumed into a smaller number of alarm intervals with durations
related to the `GASNET_EXITTIMEOUT` family of environment variables.
This ensures the user has the means to increase the intervals, and
thus addresses the original problem reported as bug 4360.

2. An "improvement" to the leader-election algorithm used for
non-collective exits has been backed-out.  The idea was that if
communication to proc0 were to timeout, one could proceed to use proc1
as the coordinator (and then proc2 if proc1 appears unreachable,
etc.).  One problem with this approach is that unless such timeouts
were to occur *uniformly*, different processes may chose different
coordinators.  Additionally, each attempt to contact a new process
restarts the timeout interval (linear in job size), resulting in worst
case time to declare the coordination attempt a failure which is
*quadratic* in the job size.
The README has also been updated to remove an (already out-of-date)
description of this difference from the ibv-conduit logic upon which
was intended to improve.

3. Prior to this commit, even if the leader-election logic for the
coordination of a non-collective did succeed.  The exit logic would
still have called the spawner's `Abort()` function.  Now the success
of the non-collective coordination leads to a `Fini()` call, as in
other conduits such as ibv.

4. Added a conduit-specific fatal signal callback to set a
`gasnetc_exit_in_signal` flag used to skip draining of the send and
receive queues when exit is (potentially) running in signal context.
This avoids one class of "messy" exit seen in my recent testing.

5. Commented out logic to free a mutex, due to it triggering an
assertion failure when the mutex is held.  This eliminates one
class of "messy" exit seen in my recent testing (at the trivial
expense of leaking a single mutex, just prior to process exit).

6. Increased `GASNETC_DEFAULT_EXITTIMEOUT_MAX` to 8 minutes, based on
recent `MPI_Finalize()` timings on Summit, extrapolated (roughly) to
full system scale.