ibv: exit improvements (including for bug 4360)
This commit makes the following improvements to the logic implementing
`gasnet_exit()` in ibv-conduit:

1. All fixed-duration `alarm()` calls in the exit code have been
subsumed into a smaller number of alarm intervals with durations
related to the `GASNET_EXITTIMEOUT` family of environment variables.
This ensures the user has the means to increase the intervals, and
thus addresses the original problem reported as bug 4360.

2. An improvement to the leader-election algorithm introduced in
ucx-conduit has been adopted into ibv-conduit.  Specifically,
`gasnetc_get_exit_role()` now enforces a timeout, which is two seconds
shorter than the surrounding alarm interval.  This allows for
signal-free detection of the inability to coordinate a non-collective
exit.  The net result is that the `testexit` cases 5 and 9 no longer
exit with the following warning (potentially from multiple processes):
    `gasnet_exit(): WARNING: timeout during exit... goodbye.`

3. Increased `GASNETC_DEFAULT_EXITTIMEOUT_MAX` to 8 minutes, based on
recent `MPI_Finalize()` timings on Summit, extrapolated (roughly) to
full system scale.