ibv,ucx: improved exit timeout reporting
This commit makes identical improvements to the handling of alarm
timer expiration in ibv- and ucx-conduits.

1. Tracking of the "exit state" (the step in progress at timer
expiration) and its appearance in a message at alarm expiration is no
longer exclusive to a DEBUG build.

2. Use of `strlen()` in constructing a `write()` call has been
replaced with `gasneti_strnlen()` to reduce the chances of "bad
behavior" if the pointer to the exit state string is corrupted.

3. In ibv-conduit only, a call to `gasneti_close_streams()` has been
removed from the exit logic to allow the alarm expiration message to
actually appear.  This was the only conduit with such a call.

Note that we already have an open bug for factoring the exit handling
logic.