Fix warning suppression for testcxx_mismatch
The recent addition of `GASNETT_USE_PRAGMA_GCC_DIAGNOSTIC`, and its
use in `testgasnet.c`, lead to the reappearance of harmless sequence
point warnings when compiling `testcxx` with `g++` and
`GASNETT_COMPILER_FORCE_MISMATCH` set.  This commit asserts the new
`GASNETT_USE_...` for this test with GNU compilers, in order to
selectively silence the warnings.