ofi: update to 1.5 API for mr_mode
This commit resolves bug 4374:
    Update mr_mode handling to OFI 1.5 API spec (or newer?)

If the header reports a version 1.5 or higher, then we request the API
1.5 when calling `fi_getinfo()` and use the fine-grained `mr_mode`
bits.  Otherwise, we continue to request the 1.0 API, including use of
the `FI_MR_BASIC` and `FI_MR_SCALABLE` constants for `mr_mode`.

The conduit README has been updated to match.

Additionally, the `tcp;ofi_rxm` provider's configure-time setting of
`GASNETC_OFI_HAS_MR_SCALABLE` is updated to `1`.  It turns out that
use of the newer API resolves the unexplained lack of `FI_MR_SCALABLE`
support in this provider (which was noted in comment 4 of bug 3867).