ofi: work around bug 4376 (EAGAIN from fi_recvmsg)
This commit adds logic to handle the scenario, detailed in bug 4376,
in which a reposting of a recycled multi-recv buffer (used for AM
reception) returns `-FI_EAGAIN`.  This behavior has been observed with
"cxi" and "udp;ofi_rxd" providers.

The additional logic adds four branches with static prediction hints
to the normal case critical path in AMPoll.  Since this is not a
behavior common to all providers, the compilation of the new logic is
conditional, with defaults and an override in the configure script.