ofi: add cxi to high-performance provider list
This commit modifies ofi-conduit to consider "cxi" to be a
high-performance provider.  This leads to suppression of the startup
message about lack of performance validation associated with most
other providers.  It does *not* suppress the message about the
experimental status of the conduit.

As with "psm2" for Intel OmniPath, the "cxi" provider is the best
support GASNet-EX has for the HPE Slingshot-11 network.  Thus these
two alone comprise the list of "high-performance providers".