mpi,udp: tuning single-nbrhd AMPoll
This commit modifies the implementation of `gasnet_AMPoll()` in both
mpi- and udp-conduits to (a) avoid taking the lock more frequently
than necessary in a single-nbrhd run and (b) reduce the length of the
lock critical section when it is taken.

In the case of mpi-conduit, the pre-existing logic to make an actual
call into AMMPI for only 1 gasnet-level call in 256 has been modified
to perform the conditional *without* holding any lock.  This is safe
because there is no reason the 1/256 ratio must be maintained exactly.

In the case of udp-conduit, there are two distinct calls to be made
depending on the number of nbrhds.  As with mpi-conduit commit, we've
moved the conditional logic outside of the lock critical section.
Additionally, the logic for single-nbrhd runs now includes the same
1-in-256 condition as in mpi-conduit.

In both conduits, the length of the critical path for multi-nbrhd
should be unchanged while the average length for single-nbhrd should
now be shorter.  Additionally, the lock critical sections are now
shorter in all cases for both conduits.