configure: Refine -with-hip-platform=nvidia support
Stop hard-coding a dependence on libcudart (which we've seen cause
problems with PGI that has a different spelling for the same
functionality).

Also since we default to -lcudart, add -L$CUDA_HOME/lib64 to the
HIP_LDFLAGS default, as it's required in general but might be
absent from CUDA_LDFLAGS (which only looked for libcuda). This
situation arose on summit using the system rocm-hip/4.3.0 module.