ofi: native implementation of extended API
This implementation is a simple forward-port of the extended API
implementation of the ofi-conduit in GASNet-1.  In particular there
are only "bulk" and "non-bulk" (`DEFER` and `NOW`) behaviors available
for `lc_opt`, with the others mapped to `NOW`.

Until now, the conduit-specific RDMA put and get logic from GASNet-1
has remained unreachable in `gasnet_ofi.c` and underwent only minor
updates for changes to types.  So, this commit makes additional
updates to resolve latent bugs that were not found by compiler
type checks or the implementation of AM Long.

1. Updated completion handling for NBI RMA operations to correctly
interact with NBI access regions (aop) and their offset-by-one
counters.

2. Updated the handling of memory registration keys for non-bulk
(EVENT_NOW) Puts to accommodate the split of the aux and client
segments.

3. Repaired logic for non-bulk (EVENT_NOW) Puts to use
`gasneti_wait()` on an internal EOP.  This resolves some efficiency
issues (such as gross over synchronization in the NBI case), while
simultaneously addressing a correctness issue (hang) with aops.

4. Removed some dead code.

This implantation of the extended API has uncovered two issues which
are not actually within implementation of RMA, but are exposed or
created as a result of the switch from the AM reference
implementation.  These will be addressed in additional commits to
follow.