testsegment: allow for single-segment conduits
This commit weakens `testsegment` for conduits (ofi in particular)
which support only a *single* call to either `gex_Segment_Attach()`
or `gex_Segment_Create()`.  In the absence of multi-segment support,
the test continues to default to testing use of a client-allocated
segment (currently the only such coverage involving host memory).

The bulk of this commit is changes to indentation.