ofi: factor FI_ environment variable handling
This commit factors the logic used to conditionally set the
environment variable `FI_PSM2_LAZY_CONN`.

Additionally, this commit adds a (disabled for now, pending bug 4413)
use of the factored logic to set `FI_UNIVERSE_SIZE`.

This same approach is anticipated to be used for other
conduit-specific environment settings in the future.