ofi/cxi: enable hybrid rx match mode by default
This commit enables "hybrid" receive matching mode for the cxi
provider (but only if there is not already an explicit setting).

Use of hybrid mode increases the stability of the conduit in the
presence of intense in-bound AM traffic without reducing the
performance of the common case.  Specifically, if the buffers posted
for offloaded matching are exhausted, then a software-based mechanism
is triggered to prevent fatal errors.  When sufficient resources are
available, offloaded matching is still used.

Current settings for protocol switch points have been tested for
reliability, but not tuned for the performance of the (uncommon)
degraded case.  The settings have been chosen to favor rendezvous over
eager, under the assumption that eager buffering of a "storm" of
in-bound AMs can lead to unwanted memory pressure.

The environment variables set in this commit are documented in the
`fi_cxi` manpage.