ofi: process multiple events per poll
This commit revises the logic for AM receive polling to process
multiple events per call.

Previously there was at most one Request and one Reply processed per
call, including the one made at the start of a Request injection.
Under certain conditions in which back-pressure did not lead to
spin-polling, this would allow full queues in the provider to remain
full as steady state.  In turn this could lead to additional buffering
of eager protocols, or overheads of rendezvous protocols.