Address bug 4361 using TreePutSeg for ReduceToAll
This commit addresses bug 4361 "RFE: handle reductions on DT_USER of
unbounded length" by increasing the max supported `dt_sz` for
ReduceToAll to approximately 32KB, matching the documentation in
`docs/GASNet-EX.txt`.  This accomplished using the newly generalized
TreePutSeg algorithm.

Note that the bug is an RFE for *unbounded* size, which this does NOT
provide.  However, 32KB is hopefully large enough for most purposes.