Add logic to free GASNet-allocated host segments
This commit adds data and code to track whether segment memory
was allocated by the client vs GASNet, as well as logic to free
GASNet-allocated segments in `gasneti_segmentDestroy()`.

Note that `gasneti_segmentDestroy()` is NOT yet reachable other
than when failing to create a segment.  However, it will be the
basis of of a complete (near future) implementation of
`gex_Segment_Destroy()`.