ucx: implement a Segment Destroy hook
This commit implements `gasnetc_segment_destroy_hook()` for
ucx-conduit *cloning* the exit-time shutdown equivalent to create
`gasnetc_segment_deregister()`.

Due to the current approach to management of registration keys, it is
non-trivial to factor (as opposed to clone) the logic.  However, a
TODO comment has been left to encourage that cleanup in the future.