fixup: ofi: update layout of gasnetc_ofi_ctxt_t
Prior work to fix the layout of `gasnetc_ofi_ctxt_t` broke support for
psm2 provider.  Specifically, the handling of multi-receive buffers
was subtly broken leading to AM-intensive tests eventually hanging.
This problem was (so far) unique to only psm2 provider, which may be
the only one making use of the `struct fi_context` field in
`gasnetc_ofi_ctxt_t`.

There are multiple conduit data structures which include a `struct
fi_context` as a member.  The conduit must pass the address of such a
structure to certain calls, or include its address in another data
structure.  The address is then included in cq entries.  The conduit
code is free to put anything it wants after (or before) the `struct
fi_context`, but that space is reserved to the provider.  The commit
which this one repairs broke that contract when reordering the fields
in `gasnetc_ofi_ctxt_t`.

This commit restores the `struct fi_context` field's place as the
first member of `gasnetc_ofi_ctxt_t` (and observes that it never
needed to lose that position, since it is always a multiple of 8-bytes
in length).

Additionally, comments are added to the relevant struct definitions to
prevent a similar mistake in the future.  The alternative of making
proper use of `gasneti_container_of()` does not seem worth the effort.