ofi: set conservative default barrier algorithm
This commit sets the default barrier algorithm for ofi-conduit to
"AMDISSEM", the same as prior to the addition of a native extended API
implementation.

This is a conservative work-around for Bug 4427 (ofi-conduit failures
with RDMADISSEM barrier).  By "conservative" I mean that it assumes no
provider is immune to the root cause, which has yet to be identified.