ofi: fix locking error in OFI_RETRY_RECVMSG logic
This commit corrects a locking bug in the handling of `EAGAIN` from
`fi_recvmsg()`.  The issue is that for thread-safe providers (all but
psm2), the macro `GASNETC_OFI_LOCK()` is a no-op.  Therefore, its use
to protect the data structures managing buffers for deferred reposting
was erroneous.  The solution applied in this commit is to use
`GASNETC_OFI_PAR_LOCK()` (note the `PAR_`) with the same lock used
for the other manipulations of the buffers.