ofi: parse env after provider and device selection
This commit delays the parse of most `GASNET_OFI_*` environment
variables until after the provider and device (aka "domain") have been
determined.  The parsing of `GASNET_OFI_DEVICE` is the sole exception,
since it influences the selection of provider and device.

This enables future work to introduce provider-specific (such as
"psm2" vs "cxi") defaults for environment settings, as well as more
fine-grained defaults based on domain (such as "mlx5*" vs "mlx4*" with
the verbs provider).