Disable ofi-conduit by default on Cray XC
This commit removes an explicit `--enable-ofi=probe` from the
cross-configure scripts for the Cray XC series.  Due to the presence
of `--disable-auto-conduit-detect`, this is essentially a disable (as
is also the case for ibv and ucx conduits).

Since an end-user is still welcome to append `--enable-ofi`, we retain
`--with-ofi-provider=gni`.