gasneti_console_messageVA: Add an optional rank argument
The new console_procid argument, when non-negative, uniquely identifies
the rank of the process who should send the message to the console with no
procid prefix. The intended use is for "job-wide" messages that are
issued collectively and/or by a single rank on behalf of the entire job.

Output is still written to the tracefile of all callers who have tracing
enabled.

Update all current callers to pass -1 for the new argument, thus
skipping the new functionality for no net change in behavior.