ofi: address provider selection issues (bug 4432)
This commit resolves the two issues raised in bug 4432:

+ Passing the semi-colon used in `verbs;ofi_rxm` and others is fragile.
+ Passing `--with-ofi-provider=arbitrary_string` warns and then
  enables the entire "known_providers" list unconditionally.

Specifically, this commit implements the following new behaviors with
respect to the argument to `--with-ofi-provider=`:

+ Now accepts `verbs`, `tcp` and `udp` as aliases for the value with
  the appropriate `;ofi_rxm` or `;ofi_rxd` appended.
+ Now converts colons in the provider string to semi-colons.
+ By default, if the provider name is unrecognized (after the aliasing
  described in the two preceding bullets) then an error is issued.
+ The behavior in the previous bullet can be overridden by passing the
  new option `--enable-allow-unknown-ofi-provider`.  This is intended
  to be used only by developers trying to support new providers, and
  is therefore undocumented.

The two aliasing behaviors are documented in `ofi-conduit/README`.

Additionally, the warning and error messages have been cleaned up to
better (IMO) convey what is happening.

Item "1c" of bug 4432 suggests `--with-ofi-provider="prov1 prov2"` as
a possible syntax for an allow-list behavior.  As mentioned in the
bug report, that would likely motivate non-trivial logic to identify
the union of the providers' features.  Even in the absence of that
optimization, such an allow-list would require non-trivial changes to
configure, for instance a loop over the words in the value to apply
the new behaviors introduced in this commit.  So, the "1c" allow-list
behavior is NOT implemented in this commit.