ofi: remove indirect fn call from AM poll path
This commit simplifies the recycling of AM send buffers by eliminating
an indirect function call to a pair of single-statement functions.
Instead a fixed function is used, which additionally provides a
logical place to implement async LC completion logic in the future.
In anticipation of that, the call sites related to "early retirement"
of AM buffers sent using `fi_inject()` have been replaced with just
the single statement needed to free the buffer.