ofi: repair logic to cancel AM multi-recv ops
This commit corrects a logic error in which a structure was
inadvertently copied.  The address of a field in the copy was then
passed as the operation context to `fi_cancel()`.  Since this did not
match the address of the original, the operation failed (but we ignore
the failure).

Additionally, general cleanups have been applied to both the posting
and cancellation of the multi-recv buffers.