ofi: revise context types
This commit revises the use of conduit-specific types passed as
operation context to libfabric.  The main changes are:

+ No longer make any implicit layout assumptions, such as that a
  `struct fi_context` or `rdma_callback_fn` is first in a structure.
  Consequently, all instances of the following have been removed:
     `// Conduit code assumes ctxt is the first field`
     `// Conduit code assumes callback is the first field`
  Those fields may now be freely moved within the structures (but
  see next item)
+ There is one remaining restriction/caveat regarding layout of some
  struct types, which is documented by two instances of:
     `// NOTE: first sizeof(void*) is overwritten when on freelist`
  This means that if one does move `ctxt` or `callback` around in
  struct definitions, one must take care what ends up in the first
  `sizeof(void*)` bytes.
+ Rename `rdma_callback_fn` to `gasnetc_rdma_callback_fn`, since its
  definition in `gasnet_core_internal.h` means it appears in sources
  outside the conduit.
+ Rename `gasnetc_ofi_ctxt_t` to `gasnetc_ofi_recv_ctxt_t` to better
  reflect its role.
+ Rename `gasnetc_ofi_am_buf_t` to `gasnetc_ofi_send_ctxt_t` to better
  reflect its role.

There are no changes to any structure layouts in this commit.
Consequently, all addresses passed as "contexts" are unchanged.
However, this work makes way for future changes.

One such change would be moving the `struct fi_context` to the *end*
of the types which embed it.  Doing so will allow omitting its
allocation for providers which do not set `FI_CONTEXT` in the mode
bits, and doubling it for providers which set `FI_CONTEXT2` in the
mode bits.

Another change enabled by this work would be relocation of the
`callback` fields away from the first position (where freelist linkage
clobbers them) to allow them to be set once at allocation rather than
at each and every use.