ofi: persistent callback functions
This commit removes (re)initialization of the `callback` field in the
non-blocking RMA contexts ("nb" and "bounce") from the RMA injection
paths.  Instead the `callback` is set when the context structure is
allocated.  The prior assignment statements have been replaced by
assertions.

In the case of `gasnetc_ofi_bounce_op_ctxt_t` this change required
reordering fields to prevent `callback` from being overwritten when
the struct is on a freelist.